+++ src/components/DefineCompany/index.ts
@@ -1,0 +1,5 @@
+import BusinessContactInfo from './BusinessContactInfo.vue'
+
+export {
+  BusinessContactInfo
+}
