+++ src/components/ReviewConfirm/Certify.vue
@@ -42,8 +42,11 @@
 </template>
 
 <script lang="ts">
+// Libraries
 import { Component, Vue, Prop, Emit } from 'vue-property-decorator'
 import { State } from 'vuex-class'
+
+// Interfaces
 import { CertifyStatementIF } from '@/interfaces'
 
 @Component({})
@@ -55,13 +58,14 @@ export default class Certify extends Vue {
   @Prop({ default: '' })
   private certifiedBy: string
 
-  private isCertified: boolean = false
-
   @Prop({ default: () => {
     return {}
   }
   })
   private certifyStatementResource: CertifyStatementIF
+
+  // Properties
+  private isCertified: boolean = false
 
   /**
    * Lifecycle callback to always give the parent a "valid" event for its property values.
