+++ src/components/Summary/index.ts
@@ -1,0 +1,5 @@
+import SummaryDefineCompany from './SummaryDefineCompany.vue'
+
+export {
+  SummaryDefineCompany
+}
