+++ src/interfaces/index.ts
@@ -8,8 +8,13 @@ export * from './store-interfaces/action-interface'
 export * from './store-interfaces/tombstone-interface'
 export * from './store-interfaces/getter-interface'
 
+// Stepper Interfaces
+export * from './stepper-interfaces/DefineCompany/business-contact-interface'
+export * from './stepper-interfaces/DefineCompany/define-company-interface'
+export * from './stepper-interfaces/ReviewConfirm/certify-interface'
+export * from './stepper-interfaces/ReviewConfirm/certify-interface'
+
 // Others
 export * from './filingData-interface'
 export * from './resource-interfaces/component-resource-interfaces/certifyStatement-interface'
-export * from './stepper-interfaces/ReviewConfirm/certify-interface'
 export * from './resource-interfaces/ExternalResourceIF'
